feat(write): only outbound write direct to db
Refs: #1923
