+++ .github/workflows/_deploy-cdk.yml
@@ -211,6 +211,33 @@ jobs:
           AWS_DEFAULT_REGION: ${{ inputs.AWS_REGION }}
           METRIPORT_VERSION: ${{ github.sha }}
 
+      # API Stack
+      - name: Diff API Stack CDK
+        uses: metriport/deploy-with-cdk@master
+        with:
+          cdk_action: "diff"
+          cdk_version: "2.122.0"
+          cdk_stack: "${{ inputs.cdk_stack }}"
+          cdk_env: "${{ inputs.deploy_env }}"
+        env:
+          INPUT_PATH: "metriport/packages/infra"
+          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
+          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+          AWS_DEFAULT_REGION: ${{ inputs.AWS_REGION }}
+      - name: Deploy API Stack CDK
+        uses: metriport/deploy-with-cdk@master
+        with:
+          cdk_action: "deploy --verbose --require-approval never"
+          cdk_version: "2.122.0"
+          cdk_stack: "${{ inputs.cdk_stack }}"
+          cdk_env: "${{ inputs.deploy_env }}"
+        env:
+          INPUT_PATH: "metriport/packages/infra"
+          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
+          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+          AWS_DEFAULT_REGION: ${{ inputs.AWS_REGION }}
+          METRIPORT_VERSION: ${{ github.sha }}
+
       # IHE Stack 
       - name: Diff IHE Stack CDK
         uses: metriport/deploy-with-cdk@master
@@ -232,33 +259,6 @@ jobs:
           cdk_action: "deploy --verbose --require-approval never"
           cdk_version: "2.122.0"
           cdk_stack: "${{ inputs.ihe_stack }}"
-          cdk_env: "${{ inputs.deploy_env }}"
-        env:
-          INPUT_PATH: "metriport/packages/infra"
-          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
-          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-          AWS_DEFAULT_REGION: ${{ inputs.AWS_REGION }}
-          METRIPORT_VERSION: ${{ github.sha }}
-
-      # API Stack
-      - name: Diff API Stack CDK
-        uses: metriport/deploy-with-cdk@master
-        with:
-          cdk_action: "diff"
-          cdk_version: "2.122.0"
-          cdk_stack: "${{ inputs.cdk_stack }}"
-          cdk_env: "${{ inputs.deploy_env }}"
-        env:
-          INPUT_PATH: "metriport/packages/infra"
-          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
-          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-          AWS_DEFAULT_REGION: ${{ inputs.AWS_REGION }}
-      - name: Deploy API Stack CDK
-        uses: metriport/deploy-with-cdk@master
-        with:
-          cdk_action: "deploy --verbose --require-approval never"
-          cdk_version: "2.122.0"
-          cdk_stack: "${{ inputs.cdk_stack }}"
           cdk_env: "${{ inputs.deploy_env }}"
         env:
           INPUT_PATH: "metriport/packages/infra"
